Move shared.irs into container mod
