+++ cmd/soroban-cli/src/commands/network/container/logs.rs
@@ -1,7 +1,7 @@
 use chrono::offset::Local;
 use futures_util::TryStreamExt;
 
-use crate::commands::network::shared::{
+use crate::commands::network::container::shared::{
     connect_to_docker, Error as ConnectionError, Network, DOCKER_HOST_HELP,
 };
