+++ cmd/soroban-cli/src/commands/network/container/shared.rs
