+++ cmd/soroban-cli/src/commands/network/container/stop.rs
@@ -1,4 +1,4 @@
-use crate::commands::network::shared::{
+use crate::commands::network::container::shared::{
     connect_to_docker, Error as ConnectionError, Network, DOCKER_HOST_HELP,
 };
